requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
botocore==1.33.6
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-debug-toolbar==4.2.0
django-graphiql-strawberry-debug-toolbar==0.2.0
frozenlist==1.4.0
gql==3.4.1
graphql-core==3.2.3
idna==3.6
jmespath==1.0.1
multidict==6.0.4
python-dateutil==2.8.2
requests==2.31.0
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.4
strawberry-graphql==0.217.0
strawberry-graphql-django==0.28.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
websockets==10.4
yarl==1.9.3
social-auth-app-django==5.4.0
django-graphql-auth==0.3.16
graphql-core==3.2.3